% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_brl}
\alias{format_brl}
\title{Format Brazilian Currency}
\usage{
format_brl(x, symbol = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{symbol}{Include R$ symbol}
}
\value{
Formatted character vector
}
\description{
Format numbers as Brazilian Real currency
}
\seealso{
Other utilities: 
\code{\link{check_insper_fonts}()},
\code{\link{format_num_br}()},
\code{\link{format_percent_br}()},
\code{\link{import_insper_fonts}()},
\code{\link{insper_caption}()},
\code{\link{save_insper_plot}()},
\code{\link{setup_insper_fonts}()},
\code{\link{use_ragg_device}()}
}
\concept{utilities}
