% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insper_palette.R
\name{insper_pal}
\alias{insper_pal}
\title{Get Insper Color Palette}
\usage{
insper_pal(palette = "primary", n = NULL, type = "discrete", reverse = FALSE)
}
\arguments{
\item{palette}{Character string indicating palette name}

\item{n}{Number of colors to return}

\item{type}{Type of palette: "discrete" or "continuous"}

\item{reverse}{Logical indicating whether to reverse palette}
}
\value{
Vector of hex color codes
}
\description{
Get Insper Color Palette
}
\examples{
insper_pal("main")
insper_pal("reds", n = 5)
}
\seealso{
\code{\link{insper_col}}, \code{\link{show_insper_palette}}, \code{\link{scale_color_insper}}, \code{\link{scale_fill_insper}}

Other colors: 
\code{\link{insper_col}()},
\code{\link{show_insper_palette}()}
}
\concept{colors}
