% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_insper_palette}
\alias{show_insper_palette}
\title{Visualize Insper Color Palettes}
\usage{
show_insper_palette(palette = "main")
}
\arguments{
\item{palette}{Character. Name of palette. Use \code{\link{list_palettes}}
to see available options. Default "main".}
}
\value{
A ggplot2 object
}
\description{
Display color palettes used by scale_*_insper_*() functions.
To see individual colors, use \code{\link{show_insper_colors}}.
}
\details{
This function visualizes the palettes that are used by ggplot2 scale functions
like \code{\link{scale_fill_insper_d}} and \code{\link{scale_color_insper_c}}.

Available palettes: main, reds, oranges, teals, grays, red_teal, red_teal_ext,
diverging, bright, contrast, categorical, accent_red, accent_teal, and more.
}
\examples{
# Show default palette
show_insper_palette()

# Show specific palettes
show_insper_palette("reds")
show_insper_palette("red_teal")
}
\seealso{
\code{\link{list_palettes}}, \code{\link{insper_pal}}, \code{\link{show_insper_colors}}

Other colors: 
\code{\link{get_insper_colors}()},
\code{\link{get_palette_colors}()},
\code{\link{insper_pal}()},
\code{\link{list_palettes}()},
\code{\link{show_insper_colors}()}
}
\concept{colors}
