% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette-utils.R
\name{show_palette_types}
\alias{show_palette_types}
\title{Show All Palette Types with Examples}
\usage{
show_palette_types(save_plot = FALSE, filename = "insper_palettes.png")
}
\arguments{
\item{save_plot}{Logical. If TRUE, saves the plot to a file. Default is FALSE.}

\item{filename}{Character. Filename for saving the plot. Only used if
`save_plot = TRUE`. Default is "insper_palettes.png".}
}
\value{
A ggplot2 object showing all palette types.
}
\description{
Creates a visual display of all Insper palette types (sequential, diverging,
and qualitative) with example plots showing recommended use cases.
}
\examples{
\dontrun{
# Display all palette types
show_palette_types()

# Save to file
show_palette_types(save_plot = TRUE, filename = "my_palettes.png")
}
}
\seealso{
Other colors: 
\code{\link{get_insper_colors}()},
\code{\link{insper_col}()},
\code{\link{insper_pal}()},
\code{\link{list_palettes}()},
\code{\link{show_insper_palette}()}
}
\concept{colors}
