% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_num_br}
\alias{format_num_br}
\title{Format Brazilian Numbers}
\usage{
format_num_br(x, digits = NULL)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{Number of decimal places}
}
\value{
Formatted character vector
}
\description{
Format numbers in Brazilian style with decimal comma and thousand separator
}
\seealso{
Other utilities: 
\code{\link{check_insper_fonts}()},
\code{\link{format_brl}()},
\code{\link{format_percent_br}()},
\code{\link{import_insper_fonts}()},
\code{\link{insper_caption}()},
\code{\link{save_insper_plot}()},
\code{\link{setup_insper_fonts}()},
\code{\link{use_ragg_device}()}
}
\concept{utilities}
