% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_percent_br}
\alias{format_percent_br}
\title{Format Brazilian Percentage}
\usage{
format_percent_br(x, digits = 1)
}
\arguments{
\item{x}{Numeric vector (proportion, not percentage)}

\item{digits}{Number of decimal places}
}
\value{
Formatted character vector
}
\description{
Format numbers as Brazilian-style percentages
}
\seealso{
Other utilities: 
\code{\link{check_insper_fonts}()},
\code{\link{format_brl}()},
\code{\link{format_num_br}()},
\code{\link{import_insper_fonts}()},
\code{\link{insper_caption}()},
\code{\link{save_insper_plot}()},
\code{\link{setup_insper_fonts}()},
\code{\link{use_ragg_device}()}
}
\concept{utilities}
