% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fossil_fuel}
\alias{fossil_fuel}
\title{Global Fossil Fuel Consumption}
\format{
A data frame with 228 rows and 5 variables:
\describe{
  \item{entity}{Character, name of the country or region (currently "World")}
  \item{code}{Character, country/region code (OWID_WRL for World)}
  \item{year}{Numeric, year of observation (1800-present)}
  \item{fuel}{Ordered factor with 3 levels: Oil, Gas, Coal.
    Levels are ordered for logical stacking in plots}
  \item{consumption}{Numeric, primary energy consumption in terawatt-hours
    (TWh)}
}
}
\source{
Our World in Data (OWID)
  \url{https://ourworldindata.org/grapher/global-fossil-fuel-consumption}
  License: CC BY 4.0
}
\usage{
fossil_fuel
}
\description{
Primary energy consumption from fossil fuels (coal, oil, and gas) measured
in terawatt-hours (TWh). Data covers global consumption from 1800 to recent years.
}
\details{
Global fossil fuel consumption since the Industrial Revolution, split by
fuel. The \code{fuel} factor is ordered Oil, Gas, Coal, which fixes the
stacking order in area charts.
}
\seealso{
\code{\link{insper_area}} for creating area plots with this data
}
\keyword{datasets}
