% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{insper_boxplot}
\alias{insper_boxplot}
\title{Insper Box Plot}
\usage{
insper_boxplot(
  data,
  x,
  y,
  fill = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{x}{Variable for x-axis}

\item{y}{Variable for y-axis}

\item{fill}{Variable for fill aesthetic}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{caption}{Plot caption}
}
\value{
ggplot object
}
\description{
Box plots with statistical annotations
}
\seealso{
\code{\link{theme_insper}}, \code{\link{scale_fill_insper}}

Other plots: 
\code{\link{insper_barplot}()},
\code{\link{insper_scatterplot}()},
\code{\link{insper_timeseries}()}
}
\concept{plots}
