% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{insper_col}
\alias{insper_col}
\title{Extract Insper Colors}
\usage{
insper_col(...)
}
\arguments{
\item{...}{Character names of colors. If none provided, returns all colors.}
}
\value{
Named character vector of hex codes
}
\description{
Extract hex codes from the Insper color palette
}
\examples{
insper_col()
insper_col("reds1", "teals1")
}
\seealso{
\code{\link{insper_pal}}, \code{\link{show_insper_palette}}

Other colors: 
\code{\link{insper_pal}()},
\code{\link{show_insper_palette}()}
}
\concept{colors}
