% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{insper_heatmap}
\alias{insper_heatmap}
\title{Insper Heatmap}
\usage{
insper_heatmap(
  data,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  show_values = TRUE
)
}
\arguments{
\item{data}{Data frame or correlation matrix}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{caption}{Plot caption}

\item{show_values}{Show correlation values on tiles}
}
\value{
ggplot object
}
\description{
Correlation matrices and heatmaps
}
