% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{insper_lollipop}
\alias{insper_lollipop}
\title{Insper Lollipop Plot}
\usage{
insper_lollipop(
  data,
  x,
  y,
  color = NULL,
  sorted = FALSE,
  point_size = 4,
  line_width = 1
)
}
\arguments{
\item{data}{A data frame containing the data to plot}

\item{x}{<[`data-masked`][ggplot2::aes_eval]> Variable for x-axis (categorical)}

\item{y}{<[`data-masked`][ggplot2::aes_eval]> Variable for y-axis (numeric)}

\item{color}{<[`data-masked`][ggplot2::aes_eval]> Variable for color aesthetic (optional)}

\item{sorted}{Logical. If TRUE, sorts by y value. Default is FALSE}

\item{point_size}{Numeric. Size of points. Default is 4}

\item{line_width}{Numeric. Width of segments. Default is 1}
}
\value{
A ggplot2 object
}
\description{
Create lollipop charts (combination of geom_segment and geom_point) using
Insper's visual identity. Ideal for displaying ranked categorical data.
}
\examples{
\dontrun{
# Simple vertical lollipop plot
df <- data.frame(category = letters[1:10], value = runif(10, 0, 100))
insper_lollipop(df, x = category, y = value)

# Sorted by value
insper_lollipop(df, x = category, y = value, sorted = TRUE)

# With color aesthetic
df$group <- rep(c("A", "B"), 5)
insper_lollipop(df, x = category, y = value, color = group)

# For horizontal orientation, swap x and y then add coord_flip()
insper_lollipop(df, x = category, y = value) + ggplot2::coord_flip()
}

}
\seealso{
\code{\link{theme_insper}}, \code{\link{scale_color_insper_d}}

Other plots: 
\code{\link{insper_area}()},
\code{\link{insper_barplot}()},
\code{\link{insper_boxplot}()},
\code{\link{insper_density}()},
\code{\link{insper_heatmap}()},
\code{\link{insper_histogram}()},
\code{\link{insper_scatterplot}()},
\code{\link{insper_timeseries}()},
\code{\link{insper_violin}()}
}
\concept{plots}
