% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insper_palette.R
\name{insper_pal}
\alias{insper_pal}
\title{Get Insper Color Palette}
\usage{
insper_pal(palette = "main", n = NULL, type = "discrete", reverse = FALSE)
}
\arguments{
\item{palette}{Character string indicating palette name}

\item{n}{Number of colors to return}

\item{type}{Type of palette: "discrete" or "continuous"}

\item{reverse}{Logical indicating whether to reverse palette}
}
\value{
Vector of hex color codes
}
\description{
Get Insper Color Palette
}
\examples{
insper_pal("main")
insper_pal("reds", n = 5)
insper_pal("red_teal")
}
\seealso{
\code{\link{show_insper_colors}}, \code{\link{show_insper_palette}}, \code{\link{scale_color_insper_d}}, \code{\link{scale_fill_insper_d}}

Other colors: 
\code{\link{get_insper_colors}()},
\code{\link{list_palettes}()},
\code{\link{show_insper_colors}()},
\code{\link{show_insper_palette}()},
\code{\link{show_palette_types}()}
}
\concept{colors}
