% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{insper_scatterplot}
\alias{insper_scatterplot}
\title{Insper Scatter Plot}
\usage{
insper_scatterplot(
  data,
  x,
  y,
  color = NULL,
  add_smooth = TRUE,
  smooth_method = "lm",
  point_size = 2,
  point_alpha = 0.7
)
}
\arguments{
\item{data}{A data frame containing the data to plot}

\item{x}{<[`data-masked`][ggplot2::aes_eval]> Variable for x-axis}

\item{y}{<[`data-masked`][ggplot2::aes_eval]> Variable for y-axis}

\item{color}{<[`data-masked`][ggplot2::aes_eval]> Variable for color aesthetic (optional)}

\item{add_smooth}{Logical. If TRUE, adds a regression line. Default is TRUE}

\item{smooth_method}{Character. Smoothing method ("lm", "loess", "gam"). Default is "lm"}

\item{point_size}{Numeric. Size of points. Default is 2}

\item{point_alpha}{Numeric. Transparency of points (0-1). Default is 0.7}
}
\value{
A ggplot2 object
}
\description{
Create scatter plots with regression lines and confidence intervals using
Insper's visual identity.
}
\examples{
\dontrun{
# Simple scatter plot
insper_scatterplot(mtcars, x = wt, y = mpg)

# Colored by variable
insper_scatterplot(mtcars, x = wt, y = mpg, color = factor(cyl))

# Without smooth line
insper_scatterplot(mtcars, x = wt, y = mpg, add_smooth = FALSE)

# With loess smoothing
insper_scatterplot(mtcars, x = wt, y = mpg, smooth_method = "loess")
}

}
\seealso{
\code{\link{theme_insper}}, \code{\link{scale_color_insper}}

Other plots: 
\code{\link{insper_area}()},
\code{\link{insper_barplot}()},
\code{\link{insper_boxplot}()},
\code{\link{insper_heatmap}()},
\code{\link{insper_lollipop}()},
\code{\link{insper_timeseries}()},
\code{\link{insper_violin}()}
}
\concept{plots}
