% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{insper_scatterplot}
\alias{insper_scatterplot}
\title{Insper Scatter Plot}
\usage{
insper_scatterplot(
  data,
  x,
  y,
  color = NULL,
  add_smooth = TRUE,
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{x}{Variable for x-axis}

\item{y}{Variable for y-axis}

\item{color}{Variable for color aesthetic}

\item{add_smooth}{Add smooth regression line}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{caption}{Plot caption}
}
\value{
ggplot object
}
\description{
Scatter plots with regression lines and confidence intervals
}
