% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{insper_timeseries}
\alias{insper_timeseries}
\title{Insper Time Series Plot}
\usage{
insper_timeseries(
  data,
  x,
  y,
  group = NULL,
  title = NULL,
  subtitle = NULL,
  caption = NULL
)
}
\arguments{
\item{data}{Data frame}

\item{x}{Time variable}

\item{y}{Value variable}

\item{group}{Grouping variable}

\item{title}{Plot title}

\item{subtitle}{Plot subtitle}

\item{caption}{Plot caption}
}
\value{
ggplot object
}
\description{
Time series plots optimized for economic/business data
}
\seealso{
\code{\link{theme_insper}}, \code{\link{scale_color_insper}}

Other plots: 
\code{\link{insper_barplot}()},
\code{\link{insper_boxplot}()},
\code{\link{insper_scatterplot}()}
}
\concept{plots}
