% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insperplot-package.R
\docType{package}
\name{insperplot-package}
\alias{insperplot}
\alias{insperplot-package}
\title{insperplot: Insper Themed ggplot2 Extensions}
\description{
Extends ggplot2 with Insper Instituto de Ensino e Pesquisa visual identity, providing custom themes, color palettes, and specialized plotting functions for academic and institutional use. This is an unofficial package created by an Insper employee, not an official Insper product.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/viniciusreginatto/insperplot}
  \item Report bugs at \url{https://github.com/viniciusreginatto/insperplot/issues}
}

}
\author{
\strong{Maintainer}: Vinicius Reginatto \email{vinicius.reginatto@example.com}

}
\keyword{internal}
