% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rec_buslines}
\alias{rec_buslines}
\title{Recife Bus Lines Data}
\format{
A data frame with 694 rows and 4 variables:
\describe{
  \item{abbrev_company}{Character, abbreviated name of the bus company
    (e.g., "BOA" for Borborema Imperial Transportes)}
  \item{code_line}{Character, unique code identifier for the bus line}
  \item{name_company}{Character, full name of the bus company operating
    the line}
  \item{name_line}{Character, name and route description of the bus line,
    typically showing origin and destination points}
}
}
\source{
Insper - Observatório Nacional de Mobilidade Sustentável
  \url{https://dataverse.datascience.insper.edu.br}
  DOI: 10.60873/FK2/TLFP8L
}
\usage{
rec_buslines
}
\description{
Data on bus lines in the Greater Recife metropolitan region from Insper's
Observatório Nacional de Mobilidade Sustentável (National Observatory of
Sustainable Mobility).
}
\details{
Reference table of bus lines and their operating companies. It shares the
\code{code_line} column with \code{\link{rec_passengers}}. Note that
\code{code_line} repeats across rows here, since a line can appear under
more than one company.
}
\seealso{
\code{\link{rec_passengers}} for related passenger transport data
}
\keyword{datasets}
