% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_insper_plot}
\alias{save_insper_plot}
\title{Save Insper Plot}
\usage{
save_insper_plot(
  plot,
  filename,
  width = height * 1.618,
  height = 6,
  dpi = 300,
  ...
)
}
\arguments{
\item{plot}{ggplot object}

\item{filename}{File name}

\item{width}{Plot width in inches}

\item{height}{Plot height in inches}

\item{dpi}{Resolution}

\item{...}{Additional arguments passed to ggsave}
}
\description{
Enhanced ggsave with institutional defaults
}
