% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_insper}
\alias{scale_color_insper}
\alias{scale_colour_insper}
\alias{scale_fill_insper}
\title{Insper Color Scales for ggplot2}
\usage{
scale_color_insper(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_colour_insper(palette = "main", discrete = TRUE, reverse = FALSE, ...)

scale_fill_insper(palette = "main", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character string indicating palette name}

\item{discrete}{Logical indicating whether to use discrete scale}

\item{reverse}{Logical indicating whether to reverse palette}

\item{...}{Additional arguments passed to ggplot2 scale functions}
}
\value{
ggplot2 scale object
}
\description{
Insper Color Scales for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, color = factor(cyl))) +
  geom_point() +
  scale_color_insper()
}
