% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_insper_c}
\alias{scale_color_insper_c}
\alias{scale_colour_insper_c}
\alias{scale_fill_insper_c}
\title{Insper Continuous Color Scale for ggplot2}
\usage{
scale_color_insper_c(palette = "turquesa", reverse = FALSE, ...)

scale_colour_insper_c(palette = "turquesa", reverse = FALSE, ...)

scale_fill_insper_c(palette = "turquesa", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character string indicating palette name}

\item{reverse}{Logical indicating whether to reverse palette}

\item{...}{Additional arguments passed to ggplot2::scale_color_gradientn()}
}
\value{
ggplot2 scale object
}
\description{
Insper Continuous Color Scale for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, color = hp)) +
  geom_point() +
  scale_color_insper_c(palette = "turquesa")
}
\seealso{
\code{\link{theme_insper}}, \code{\link{scale_color_insper_d}}

Other scales: 
\code{\link{scale_color_insper_d}()}
}
\concept{scales}
