% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_insper_d}
\alias{scale_color_insper_d}
\alias{scale_colour_insper_d}
\alias{scale_fill_insper_d}
\title{Insper Discrete Color Scale for ggplot2}
\usage{
scale_color_insper_d(palette = "main", reverse = FALSE, ...)

scale_colour_insper_d(palette = "main", reverse = FALSE, ...)

scale_fill_insper_d(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character string indicating palette name}

\item{reverse}{Logical indicating whether to reverse palette}

\item{...}{Additional arguments passed to ggplot2::discrete_scale()}
}
\value{
ggplot2 scale object
}
\description{
Insper Discrete Color Scale for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg, color = factor(cyl))) +
  geom_point() +
  scale_color_insper_d()
}
\seealso{
\code{\link{insper_pal}}, \code{\link{theme_insper}}, \code{\link{scale_color_insper_c}}

Other scales: 
\code{\link{scale_color_insper_c}()}
}
\concept{scales}
