% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_insper_colors}
\alias{show_insper_colors}
\title{Visualize All Individual Insper Colors}
\usage{
show_insper_colors(color_family = "all")
}
\arguments{
\item{color_family}{Optional filter: "all", "reds", "oranges", "magentas",
"teals", "grays", "basic". Default "all".}
}
\value{
A ggplot2 object showing the colors
}
\description{
Display all individual named colors (reds1, teals1, etc.) as a visual grid.
To extract colors by name for use in plots, use \code{\link{get_insper_colors}} instead.
To see palettes, use \code{\link{show_insper_palette}}.
}
\details{
This function creates a visual display of individual colors available in the
Insper color system. These are atomic color units (like reds1, teals2) that
can be extracted using \code{\link{get_insper_colors}}.

For palette visualization (collections of colors used in scales), use
\code{\link{show_insper_palette}} instead.
}
\examples{
# Show all individual colors
show_insper_colors()

# Show only reds
show_insper_colors("reds")

# Show only grays
show_insper_colors("grays")
}
\seealso{
\code{\link{get_insper_colors}}, \code{\link{show_insper_palette}}, \code{\link{list_palettes}}

Other colors: 
\code{\link{get_insper_colors}()},
\code{\link{get_palette_colors}()},
\code{\link{insper_pal}()},
\code{\link{list_palettes}()},
\code{\link{show_insper_palette}()}
}
\concept{colors}
