% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_insper_palette}
\alias{show_insper_palette}
\title{Show Insper Color Palette}
\usage{
show_insper_palette(palette = "all")
}
\arguments{
\item{palette}{Character string specifying palette subset ("all", "grays", "reds", "oranges", "magentas", "teals")}
}
\value{
A ggplot2 object showing the color palette
}
\description{
Display the Insper color palette visually
}
\examples{
show_insper_palette()
show_insper_palette("reds")
}
\seealso{
\code{\link{insper_col}}, \code{\link{insper_pal}}

Other colors: 
\code{\link{get_insper_colors}()},
\code{\link{insper_pal}()},
\code{\link{list_palettes}()},
\code{\link{show_insper_colors}()},
\code{\link{show_palette_types}()}
}
\concept{colors}
